refactor: using TDD - Refactoring the code

+++ sistema_ingressos/src/main/java/com/vandv/sistema_ingressos/service/ShowCrudPadraoService.java
@@ -1,6 +1,8 @@
 package com.vandv.sistema_ingressos.service;
 
 import com.vandv.sistema_ingressos.dto.ShowPostDto;
+import com.vandv.sistema_ingressos.exception.Show.IdInvalidoException;
+import com.vandv.sistema_ingressos.exception.Show.ProdutoNullException;
 import com.vandv.sistema_ingressos.model.Show;
 import com.vandv.sistema_ingressos.repository.ShowRepository;
 import org.modelmapper.ModelMapper;
@@ -17,19 +19,26 @@ public class ShowCrudPadraoService implements ShowCrudService{
     private ModelMapper modelMapper = new ModelMapper();
     @Override
     public Show showCreate(ShowPostDto showPostDto) {
+        if (showPostDto == null) {
+            throw new ProdutoNullException();
+        }
         Show show = modelMapper.map(showPostDto, Show.class);
-        System.out.println("oiiiiii");
         return showRepository.save(show);
     }
 
     @Override
     public void showDelete(Long id) {
-        showRepository.deleteById(id);
+        if(showRepository.existsById(id)) {
+            showRepository.deleteById(id);
+        } else {
+            throw new IdInvalidoException();
+        }
+
     }
 
     @Override
     public Show showFindById(Long id) {
-        return showRepository.findById(id).orElse(null);
+        return showRepository.findById(id).orElseThrow(IdInvalidoException::new);
     }
 
     @Override
